Refactor test unitarios.

+++ src/main/java/br/com/lucas/deasfio_todolist/service/TodoService.java
@@ -2,6 +2,7 @@ package br.com.lucas.deasfio_todolist.service;
 
 import br.com.lucas.deasfio_todolist.entity.Todo;
 import br.com.lucas.deasfio_todolist.repository.TodoRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
@@ -35,6 +36,10 @@ public class TodoService {
     }
 
     public List<Todo> delete(Long id){
+        if (id == null) {
+            throw new IllegalArgumentException("ID não pode ser nulo");
+        }
+
         todoRepository.deleteById(id);
         return list();
     }
